Add tracking for bash configs

+++ toRepository.sh
@@ -1,2 +1,18 @@
-cp ~/.vimrc vim/.vimrc
-cp ~/.config/i3/config i3/config
+VIMRC=~/.vimrc 
+I3CONF=~/.config/i3/config
+BASHRC=~/.bashrc 
+
+if test -f "$VIMRC"; then
+	echo "$VIMRC found"
+	cp ~/.vimrc vim/.vimrc
+fi
+
+if test -f "$I3CONF"; then
+	echo "$I3CONF found"
+	cp ~/.config/i3/config i3/config
+fi
+
+if test -f "$BASHRC"; then
+	echo "$BASHRC found"
+	cp ~/.bashrc bash/.bashrc
+fi
